Organize webhooks handlers and expand registration
