--- app/handlers/shipit/check_suite_handler.rb
@@ -1,15 +1,0 @@
-module Shipit
-  class CheckSuiteHandler < Handler
-    params do
-      requires :check_suite do
-        requires :head_sha, String
-        requires :head_branch, String
-      end
-    end
-    def process
-      stacks.where(branch: params.check_suite.head_branch).each do |stack|
-        stack.commits.where(sha: params.check_suite.head_sha).each(&:schedule_refresh_check_runs!)
-      end
-    end
-  end
-end
