--- app/handlers/shipit/membership_handler.rb
@@ -1,41 +1,0 @@
-module Shipit
-  class MembershipHandler < Handler
-    params do
-      requires :action, String
-      requires :team do
-        requires :id, Integer
-        requires :name, String
-        requires :slug, String
-        requires :url, String
-      end
-      requires :organization do
-        requires :login, String
-      end
-      requires :member do
-        requires :login, String
-      end
-    end
-    def process
-      team = find_or_create_team!
-      member = User.find_or_create_by_login!(params.member.login)
-
-      case params.action
-      when 'added'
-        team.add_member(member)
-      when 'removed'
-        team.members.delete(member)
-      else
-        raise ArgumentError, "Don't know how to perform action: `#{action.inspect}`"
-      end
-    end
-
-    private
-
-    def find_or_create_team!
-      Team.find_or_create_by!(github_id: params.team.id) do |team|
-        team.github_team = params.team
-        team.organization = params.organization.login
-      end
-    end
-  end
-end
