--- app/handlers/shipit/push_handler.rb
@@ -1,16 +1,0 @@
-module Shipit
-  class PushHandler < Handler
-    params do
-      requires :ref
-    end
-    def process
-      stacks.where(branch: branch).each(&:sync_github)
-    end
-
-    private
-
-    def branch
-      params.ref.gsub('refs/heads/', '')
-    end
-  end
-end
