--- app/handlers/shipit/status_handler.rb
@@ -1,21 +1,0 @@
-module Shipit
-  class StatusHandler < Handler
-    params do
-      requires :sha, String
-      requires :state, String
-      accepts :description, String
-      accepts :target_url, String
-      accepts :context, String
-      accepts :created_at, String
-
-      accepts :branches, Array do
-        requires :name, String
-      end
-    end
-    def process
-      Commit.where(sha: params.sha).each do |commit|
-        commit.create_status_from_github!(params)
-      end
-    end
-  end
-end
