+++ lib/shipit/webhooks/handlers.rb
@@ -1,0 +1,12 @@
+require 'shipit/webhooks/handlers/handler'
+require 'shipit/webhooks/handlers/push_handler'
+require 'shipit/webhooks/handlers/status_handler'
+require 'shipit/webhooks/handlers/membership_handler'
+require 'shipit/webhooks/handlers/check_suite_handler'
+
+module Shpit
+  module Webhooks
+    module Handlers
+    end
+  end
+end
