+++ lib/shipit/webhooks/handlers/check_suite_handler.rb
@@ -1,0 +1,20 @@
+require 'shipit/webhooks/handlers/handler'
+module Shipit
+  module Webhooks
+    module Handlers
+      class CheckSuiteHandler < Handler
+        params do
+          requires :check_suite do
+            requires :head_sha, String
+            requires :head_branch, String
+          end
+        end
+        def process
+          stacks.where(branch: params.check_suite.head_branch).each do |stack|
+            stack.commits.where(sha: params.check_suite.head_sha).each(&:schedule_refresh_check_runs!)
+          end
+        end
+      end
+    end
+  end
+end
